固定sing-box版本，使用terraform变量代替sed，优化start.sh脚本

+++ start.sh
@@ -1,20 +1,29 @@
 #!/usr/bin/env bash
 set -eu
-
+# 注意！！！！！！！！！
 # 使用前要更改下列变量 确定文件输出位置 以及 gcp项目名
 OUTPUT_FILE="/mnt/d/Scoop/apps/sing-box/current/conf.d/04-outbounds.json"
-PROJECT="dev-2026522"
+PROJECT="lab-20260627"
 
 
 build() {
     bash ./scripts/gen-keys.sh
-    sed -i "s/OUTPUT_FILE/${OUTPUT_FILE}/g" ./localfile.tf
-    echo "project_id = ${PROJECT}" > ./terraform.tfvars
+    cat > ./terraform.tfvars << EOF
+project_id = "${PROJECT}"
+client_config_path = "${OUTPUT_FILE}"
+EOF
     terraform apply -auto-approve
 }
 
 destroy() {
-    terraform destroy -auto-approve
+    read -r -p "确定要销毁全部GCP资源吗？(Y/n): " confirm
+    if [[ ${confirm:-n} =~ ^[Yy]$ || -z ${confirm} ]];then
+        terraform destroy -auto-approve
+        rm -rf ./keys.auto.tfvars
+    else
+        echo "取消销毁"
+        exit 0
+    fi
 }
 
 main() {
@@ -26,8 +35,8 @@ main() {
         destroy
     else
         echo "使用如下格式"
-        echo "./start build"
-        echo "./start destroy"
+        echo "./start.sh build"
+        echo "./start.sh destroy"
         exit 1
     fi
 }
